Loading in the new help data files. Be sure to 'checkout' this directory next update/reboot. Also, be sure to checkout src/help next reboot.
Ok.....

I rewrote the help system to be online editable and a bit cleaner and
nicer than the previous system.  All the help system commands are
under "hcollection" which is structured similar to olc.  It should feel
fimiliar.
The "help" command is still pointing to the old help system for now.

I'll put a more in depth desc of the new system somewhere soon.
Until then, those I draft into working on the help files will have to
be instructed by me.

More to come.