Greater devils (race devil, class arch, or greater, or duke) can no longer be hit by non-magical weapons. (same as noncorporeal undead)
Symbol of pain, Stigmata and Soulless are now exclusive.
Checking for this with "HAS_SYMBOL" so that only one burning
nasty mark can be on yer head at once.

Symbol of pain now only works once.  Damage is lowered _some_ for good.
Damage is negligible for neutral and irrelivent for evil.