lit bombs no longer damaged by  other bombs ( to keep from destroying themselves )
Temp buffers used in make_corpse now terminated before being used.

specials no longer called by char_to_room and char_from_room as the users
of these two functions do not take into account the death/dismemberment
of the victims nor the failure of said departure or arrival.

We're going to have to find a sane way to implement this.
We'll either have to make char_to_room return another damn enum int or
throw an exception or something.  Either that or we'll have to call the
specs before calling char_*_room.  I'm not entirely sure which.