Creature::extract() is now private.  Replacement functions are: Creature::rent(void) Creature::cryo(void) Creature::quit(void) Creature::idle(void) Creature::die(void) Creature::remort(void) Creature::purge(bool destroy_obj)
Rest of mud modified to use new functions.  Most (but not all) Crash_*
functions have been taken out.  Mechanism for loadrooms have been changed.
There is now a loadroom, a homeroom, and a hometown.  Your loadroom is the
room you will load in.  The homeroom is the room you are configured to load
in.  Your hometown determines the room you'll load in if your loadroom and
homeroom aren't set.