Proposed change to the "require alias" prog condition.
syntax: *require alias < m | o > < vnum >

gets a list of aliases for the mob or object "vnum" and looks for an exact match or an abbreviation of the keyword in the list of aliases.

only 3 mobs currently use the "require alias" condition.  These mobs can be changed
to use the "require abbrev" condition instead.