We now have a new automated auction system.  The auction channel is now unusable by mortals, but is interplanar.
There is an auctioneer spec and a courier imp spec.
A mortal may auction an item at an auctioneer using the command
auction <item> <amount>

To bid on an item a mortal may use the command:
bid <item number> <amount>

Stats are available on an auctioned item with the command:
bidstat <item number>

Once an item is one, a courier imp will take the item to the winning
bidder anywhere in the world and return the money to the items owner.