There is now a limit of 10 characters on each account.  If an account already has more than 10 characters, they will be systematically deleted starting with characters having a rent code of new, or creating.
If there are still more than 10 characters on the account, any characters
with a rent code of quit which are less than level 25 will be deleted.

If there are STILL more than 10 characters on the account, all characters
with an index of 11 or higher will be unavailable until such time as the
player deletes some characters.