Changed exp calculation and added explore bonuses
The exp calculation has been changed to reduce the exp given by
various affects and bitflags, to more resemble the actual difficulty
that the mob will give an average player.

To counter some of this affect, a exp bonus of up to 25% will be
given for fighting a wide variety of mobiles.  Currently, the system
only remembers the last 100, though this can be easily tweaked.