Update src/olc/olc_mob.c
Changed mset for height and weight to take into account player preference of metric/imperial. For imperial on height, added 2 char arrays to handle chopping the numeric argument for height into 2 arguments, so that you can put in 8 6 to get 8foot 6inches. This also required a bit of an adjustment for the conversion rounding limitations on integers. Limits are based on the original limits for cm. 

I didn't feel mob weight needed to be so precise, so it's much simpler.