Update src/olc/olc_obj.c
Changed oset to take into account player preferences of metric/imperial measure. Modified the limits to make more sense, as the new float for object weight made the number crazy high and such. This also allows for objects to be set for decimal amounts. I didn't feel it was necessary to separate the imperial measure into ounces, as I feel a decimal pound is still understandable. Maybe in the future we'll give that a go if it is necessary.