Update src/db/db.c
It is very annoying to change a mobs level and have it change the height and weight of the mob. It doesn't make sense.

So move the mana/move point adjustments from set_physical_attribs into recalculate_based_on_level, to separate the height/weight calculations from the point calculations.