Add storeroom to vendor configuration
Vendors now have a storeroom parameter.  If the vendor has a
storeroom, it will use the contents of the room instead of its
inventory for the items to be sold.

This was done primarily so consigment shops could specify a house room
as a storeroom so that objects could be automatically saved.