Take money from mailing instead of giving it.
The postmaster would adjust a mailing player's money positively
instead of negatively, allowing players to use mail as a way to get
cash.  The way this was discovered was by a player which mailed
himself many thousands of mails, causing huge amounts of db traffic
due to username lookups.

The adjustment has been fixed, and duplicate player names are now
eliminated before the db lookup.

Also, the previous commit didn't completely remove the spirit princess
spec, and this commit finishes the job.