Overhaul skill/spell listing
The commands alterations, skills, spells, and triggers now only display
their respective categories of ability.

New commands programs, songs, and zens for the other categories.

Using any of these with a class name displays the categorized abilities
for that class.

Listing at a guildmaster now only shows the abilities offered by that
specific guildmaster and not all of them.

This addresses issue #85.